requirements.txt
gitpython
pillow